pyproject.toml
[build-system]
requires = ["setuptools>=68.2.2", "wheel>=0.41.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bezzanlabs.capsule"
dynamic = ["version"]
authors = [
  {name="Vitor Bezzan", email="vitor@bezzan.com"},
]
description = "A tidy capsule for models and pipelines, with a focus in deployment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.1.1",
    "numpy>=1.26.0",
    "scipy>=1.11.3",
    "scikit-learn~=1.3.1",
    "imbalanced-learn~=0.11.0",
    "joblib~=1.3.2",
    "dill~=0.3.8",
    "pycryptodome~=3.20.0"
]

[tool.setuptools.dynamic]
version = {attr = "bezzanlabs.capsule.__version__"}

[project.optional-dependencies]
dev = [
    "pre-commit==3.5.0",
    "mypy==1.6.0",
    "pylint==3.0.1",
    "pytest==7.3.1",
    "pytest-mock",
    "pytest-cov",
    "safety==3.1.0",
    "sphinx_autodoc_typehints==1.24.1",
    "nbsphinx==0.9.3",
    "sphinx==7.2.6",
    "sphinx_copybutton==0.5.2",
    "sphinx_rtd_theme==1.3.0",
    "recommonmark==0.7.1",
    "isort==5.12.0",
    "pandas-stubs==2.1.1.230928",
]

[project.urls]
"Homepage" = "https://github.com/vitorbezzan/capsule"
"Bug Tracker" = "https://github.com/vitorbezzan/capsule"

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report html --cov-report term-missing --cov-fail-under 0"
pythonpath = [
  "src"
]
